feat(useRcTableAdapter): rename hook and replace getData function
closes #1822
