+++ src/Table/Table.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 
 import { default as RCTable } from 'rc-table'
 
-import { UseRcTableThemeProps } from '../useRcTableTheme/helper'
-import { useRcTableTheme } from '../useRcTableTheme/useRcTableTheme'
+import { UseRcTableAdapterProps } from '../useRcTableAdapter/helper'
+import { useRcTableAdapter } from '../useRcTableAdapter/useRcTableAdapter'
 
 export type TableComponentType = <ITEM = unknown>(
-  props: UseRcTableThemeProps<ITEM>
+  props: UseRcTableAdapterProps<ITEM>
 ) => React.ReactElement | null
 
 export const Table: TableComponentType = props => {
-  const tableProps = useRcTableTheme(props)
+  const tableProps = useRcTableAdapter(props)
 
   return <RCTable {...tableProps} />
 }
